finalizando parte mandatoria

+++ srcs/ft_convert_dec_hex.c
@@ -6,7 +6,7 @@
 /*   By: caredua3 <caredua3@student.42sp.org.br>    +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2023/11/25 11:57:42 by caredua3          #+#    #+#             */
-/*   Updated: 2023/11/26 14:54:40 by caredua3         ###   ########.fr       */
+/*   Updated: 2023/11/29 19:35:21 by caredua3         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -39,9 +39,9 @@ static char	*ft_assemvly_hex(t_list *l_rest, char x)
 	return (str);
 }
 
-char	*ft_convert_dec_hex(unsigned int n, char x)
+char	*ft_convert_dec_hex(unsigned long n, char x)
 {
-	int divider;
+	unsigned long divider;
 	t_list *l_rest;
 	t_list *new_node;
 	char *str_hex;
